# docs/swagger.yaml
...
openapi: 3.0.1
info:
  title: GoJobs - Nền tảng việc làm cho mọi người
  description: GoJobs API Docs
  termsOfService: http://swagger.io/terms/
  contact:
    email: trankhoa798@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:80
...
paths:
  /web:
    get:
      tags:
        - WebAdminAPI
      summary: Đi tới trang web quản lý
      produces:
        - application/html
      responses:
        "200":
          description: "Truy cập thành công"
        "404":
          description: "Trang không được tìm thấy"
...
  /web/job:
    get:
      tags:
        - WebAdminAPI
      summary: Danh sách thông tin về Job
      security:
        - cookieAuth: []
      requestBody:
        description: Thông tin phân trang
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/pagination'
      produces:
        - application/html
      responses:
        "200":
          description: "Truy cập thành công"
        "404":
          description: "Trang không được tìm thấy"
...
components:
  schemas:
...
    pagination:
      type: object
      properties:
        pageSize:
          type: number
          default: 5
        pageNumber:
          type: number  
          default: 1
        search:
          type: string
        sort:
          type: string
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token